requirements.txt
numpy==1.19.2
pandas==1.1.5
gdown==3.13.0
s2cell==1.4.0
pyyaml==5.4.1
pyarrow==2.0.0
uvicorn==0.13.4
fastapi==0.65.1